Moved scheduler from services to separated class
