+++ email/service/src/main/java/com/senla/email/service/PasswordRestoreService.java
@@ -9,6 +9,5 @@ public interface PasswordRestoreService {
 
     String confirmPasswordRestoring(NewPasswordFormDto newPasswordFormDto, String confirmationCode);
 
-    void sendPasswordRestoreNotificationMail();
 
 }
